finished toolbar cleanup
