+++ package.json
@@ -47,6 +47,7 @@
     "react-router-dom": "^4.1.2",
     "react-transition-group": "^1.2.0",
     "reactstrap": "^4.8.0",
-    "style-loader": "^0.18.2"
+    "style-loader": "^0.18.2",
+    "underscore": "^1.8.3"
   }
 }
