--- reactApp/labels/colorLabels.js
@@ -1,9 +1,0 @@
-export default [
-        {label: 'Red', style: 'red'},
-        {label: 'Orange', style: 'orange'},
-        {label: 'Yellow', style: 'yellow'},
-        {label: 'Green', style: 'green'},
-        {label: 'Blue', style: 'blue'},
-        {label: 'Indigo', style: 'indigo'},
-        {label: 'Violet', style: 'violet'},
-];
